mcp/requirements.txt
fastapi
google-cloud-firestore
uvicorn
